Added policy based authorization for update and delete endpoints

+++ Controllers/ParksController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using NationalParksApi.Models;
 using NationalParksApi.Repositories;
 using Microsoft.Extensions.Logging;
+using SQLitePCL;
 
 namespace NationalParksApi.Controllers
 {
@@ -14,12 +15,14 @@ namespace NationalParksApi.Controllers
         private readonly IParkRepository _repo;
         private readonly IWeatherService _weatherService;
         private readonly ILogger<ParksController> _logger;
+        private readonly IAuthorizationService _authorizationService;
 
-        public ParksController(IParkRepository repo, IWeatherService weatherService, ILogger<ParksController> logger)
+        public ParksController(IParkRepository repo, IWeatherService weatherService, ILogger<ParksController> logger, IAuthorizationService authorizationService)
         {
             _repo = repo;
             _weatherService = weatherService;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
 
         [Authorize(Roles = "Admin,User")]
@@ -53,41 +56,70 @@ namespace NationalParksApi.Controllers
         }
 
         [HttpPost]
-        [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "Admin, User")]
         public IActionResult CreatePark([FromBody] NationalPark newPark)
         {
-            _logger.LogInformation("Creating new park: {Name} in {State}, established in {YearEstablished} at (Lat: {Latitude}, Lon: {Longitude})",
-            newPark.Name, newPark.State, newPark.YearEstablished, newPark.Latitude, newPark.Longitude);
+            var currentUserEmail = User.Identity?.Name;
+            if (!string.IsNullOrEmpty(currentUserEmail))
+            {
+                newPark.CreatedBy = currentUserEmail;
+            }
+
+            _logger.LogInformation("Creating new park: {Name} in {State}, established in {YearEstablished} at (Lat: {Latitude}, Lon: {Longitude}) created by {CreatedBy}.",
+            newPark.Name, newPark.State, newPark.YearEstablished, newPark.Latitude, newPark.Longitude, newPark.CreatedBy);
 
             var created = _repo.Add(newPark);
             return CreatedAtAction(nameof(GetParks), new { id = created.Id }, created);
         }
 
         [HttpPut("{id}")]
-        [Authorize(Roles = "Admin")]
-        public IActionResult UpdatePark(int id, [FromBody] NationalPark updatedPark)
+        [Authorize]
+        public async Task<IActionResult> UpdatePark(int id, [FromBody] NationalPark updatedPark)
         {
             _logger.LogInformation("Updating park: {Name} in {State}, established in {YearEstablished} at (Lat: {Latitude}, Lon: {Longitude})",
             updatedPark.Name, updatedPark.State, updatedPark.YearEstablished, updatedPark.Latitude, updatedPark.Longitude);
+            var existingPark = _repo.GetById(id);
+            if (existingPark == null)
+            {
+                _logger.LogWarning("Park with ID {Id} not found.", id);
+                return NotFound("Park not found.");
+            }
+
+            var authoriationResult = await _authorizationService.AuthorizeAsync(User, existingPark, "OwnerOrAdmin");
+            if (!authoriationResult.Succeeded)
+            {
+                _logger.LogWarning("User is not authorized to update park with ID {Id}.", id);
+                return Forbid();
+            }
+
             var park = _repo.Update(id, updatedPark);
+            return Ok(park);
+        }
+
+        [HttpDelete("{id}")]
+        [Authorize]
+        public async Task<IActionResult> DeletePark(int id)
+        {
+            _logger.LogInformation("Deleting park with ID {Id}", id);
+            var park = _repo.GetById(id);
             if (park == null)
             {
                 _logger.LogWarning("Park with ID {Id} not found.", id);
                 return NotFound("Park not found.");
             }
-            return Ok(park);
-        }
 
-        [HttpDelete("{id}")]
-        [Authorize(Roles = "Admin")]
-        public IActionResult DeletePark(int id)
-        {
-            _logger.LogInformation("Deleting park with ID {Id}", id);
+            var authoriationResult = await _authorizationService.AuthorizeAsync(User, park, "OwnerOrAdmin");
+            if (!authoriationResult.Succeeded)
+            {
+                _logger.LogWarning("User is not authorized to delete park with ID {Id}.", id);
+                return Forbid();
+            }
+            
             var success = _repo.Delete(id);
             if (!success)
             {
-                _logger.LogWarning("Park with ID {Id} not found.", id);
-                return NotFound("Park not found.");
+                _logger.LogError("Failed to delete park with ID {Id}.", id);
+                return StatusCode(StatusCodes.Status500InternalServerError, "Failed to delete park.");
             }
             return Ok($"Park with ID {id} deleted.");
         }
